simplify schemes
